% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{charite}
\alias{charite}
\title{Tinnitus patients data from Charite Universitaetsmedizin Berlin.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6355 rows and 367 columns.
}
\usage{
charite
}
\description{
Tinnitus patients data from Charite Universitaetsmedizin Berlin.
}
\keyword{datasets}
