% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dict}
\alias{data_dict}
\title{Charite; tinnitus data dictionary.}
\format{
A data frame with 5 variables:
\describe{
\item{item}{\code{chr} Name of the variable. Meta variables like ".jour_nr" start
with a ".".}
\item{label}{\code{chr} Label of the variable. Note, that while every variable has a
name, not every variable has a label.}
\item{description}{\code{chr} Description of the variable.}
\item{values}{\code{list} Mapping of numeric values to character strings for
categorical variables. \code{NULL} for numeric variables.}
...
}
}
\source{
\url{https://tinnituszentrum.charite.de/}
}
\usage{
data_dict
}
\description{
A data frame with descriptions for every variable in the Charite tinnitus
questionnaire data.
}
\details{
Note: As of 17.08.2020, english descriptions of variables of the following
questionnaires are available: meta, acsa, adsl, bi, bsf, isr, phq, psq, sozk,
sf8, tq, tinskal, tlq.

The descriptions for variables of soc and swop have yet to be translated.
}
\keyword{datasets}
