% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_treated.R
\name{filter_treated}
\alias{filter_treated}
\title{Extract the "legitimate" subset of patients who were treated}
\usage{
filter_treated(df, tolerance = 14L)
}
\arguments{
\item{df}{The \code{charite} data.}

\item{tolerance}{The tolerable maximum number of days between an A recording
and its corresponding E recording.}
}
\value{
The filtered \code{charite}dataset. Each patient has exactly 1 A and 1 E
recording.
}
\description{
In many scenarios, it is useful to extract the subset of patients who
received treatment. However, for several patients, there exist multiple
recordings in phase 'A' or 'E', respectively.
The function \code{filter_treated} finds the "closest" A recording to the first E
recording within the boundary of \code{tolerance} days.
Since the treatment was 7 days long, the function considers the A recording
that is closest to 7 days before the E recording (in case there are multiple
A recordings for the same patient).
}
